Small refactor, added Javadoc and comments.
