+++ engine-tests/src/test/java/org/terasology/replayTests/ExampleReplayTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 MovingBlocks
+ * Copyright 2017 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.recording;
+package org.terasology.replayTests;
 
-import org.junit.After;
 import org.junit.Test;
 import org.terasology.ReplayTestingEnvironment;
 import org.terasology.entitySystem.entity.EntityRef;
@@ -24,6 +23,8 @@ import org.terasology.logic.location.LocationComponent;
 import org.terasology.logic.players.LocalPlayer;
 import org.terasology.math.geom.Vector3f;
 import org.terasology.math.geom.Vector3i;
+import org.terasology.recording.EventSystemReplayImpl;
+import org.terasology.recording.RecordAndReplayStatus;
 import org.terasology.registry.CoreRegistry;
 import org.terasology.world.WorldProvider;
 
@@ -45,11 +46,6 @@ public class ExampleReplayTest extends ReplayTestingEnvironment {
         }
     };
 
-   /* @After
-    public void closeGame() throws Exception {
-        super.getHost().shutdown();
-        replayThread.join();
-    }*/
 
     @Test
     public void testExampleRecordingPlayerPosition() {
@@ -98,6 +94,7 @@ public class ExampleReplayTest extends ReplayTestingEnvironment {
             Vector3i blockLocation2 = new Vector3i(26, 13, -3);
             Vector3i blockLocation3 = new Vector3i(26, 12, -2);
 
+            //checks the block initial type of three chunks that will be modified during the replay.
             WorldProvider worldProvider = CoreRegistry.get(WorldProvider.class);
             assertEquals(worldProvider.getBlock(blockLocation1).getDisplayName(), "Grass");
             assertEquals(worldProvider.getBlock(blockLocation2).getDisplayName(), "Air");
@@ -107,6 +104,7 @@ public class ExampleReplayTest extends ReplayTestingEnvironment {
                 Thread.sleep(1000);
             }//The replay is finished at this point
 
+            //checks the same blocks again after the replay.
             assertEquals(worldProvider.getBlock(blockLocation1).getDisplayName(), "Grass");
             assertEquals(worldProvider.getBlock(blockLocation2).getDisplayName(), "Grass");
             assertEquals(worldProvider.getBlock(blockLocation3).getDisplayName(), "Air");
